movies_async_api/requirements/common.txt
fastapi==0.61.1
uvicorn==0.12.2
elasticsearch[async]==7.9.1
aioredis==1.3.1
orjson==3.4.3
python-decouple==3.4
backoff==1.10.0
